fix writing data to db incorrectly

+++ app.js
@@ -13,27 +13,32 @@ const api = require('./model/Api')
 const app = express()
 
 app.use(cookieParser())
-app.use(express.static('view'))
 app.use(bodyParser.urlencoded({ extended: true }))
 
 
-app.get('/', (req, res) => {
-    db.getLatestId()
-        .then(id => {
-            if (!req.cookies['id']) {
+app.get('/', (req, res, next) => {
+    if (!req.cookies['id']) {
+        db.getLatestId()
+            .then(id => {
+                console.log('setting new cookie')
                 res.cookie('id', id + 1)
-                res.sendFile('./view/index.html')
+                next()
                 return db.createUser(id+1)
-            }
-            res.sendFile(path.join(__dirname + '/view/index.html'))
-        })
-        .catch(err => {
-            console.log(err.message)
-        })
+            })
+            .catch(err => {
+                res.status(500)
+                console.log(err.message)
+            })
+    } else {
+        next()
+    }
+
 })
 
+app.use(express.static('view'))
+
 app.get('/weather/coordinates', ((req, res) => {
-    if(req.query.lon && req.query.lat) {
+    if(req.query.lon && req.query.lat)
         api.getCurrentWeather(null, req.query.lon, req.query.lat)
             .then(data => {
                 res.status(data.status)
@@ -43,29 +48,33 @@ app.get('/weather/coordinates', ((req, res) => {
                 res.status(500)
                 console.log(err.message)
             })
-    } else {
+    else
         res.status(404)
-    }
+
 }))
 
 app.get('/weather/city', (req, res) => {
     // console.log(req.query.q)
-    api.getCurrentWeather(req.query.q)
-        .then(data => {
-            // const weather = data.weather
-            // const main = data.main
-            // console.log(data.status)
-            res.status(data.status)
-            res.json(data.json)
+    if(req.query.q)
+        api.getCurrentWeather(req.query.q)
+            .then(data => {
+                // const weather = data.weather
+                // const main = data.main
+                // console.log(data.status)
+                res.status(data.status)
+                res.json(data.json)
 
-        })
-        .catch(err => {
-            console.log(err.message)
-        })
+            })
+            .catch(err => {
+                res.status(500)
+                console.log(err.message)
+            })
+    else
+        res.status(404)
 })
 
 app.get('/favourites/cities', (req, res) => {
-    if(req.query.userId) {
+    if(req.query.userId)
         db.getUserCities(req.query.userId)
             .then(rows => {
                 let cities = []
@@ -78,27 +87,41 @@ app.get('/favourites/cities', (req, res) => {
                 })
             })
             .catch(err => {
+                res.status(200)
                 console.log(err.message)
             })
-    } else {
+    else
         res.status(404)
-    }
 })
 
 app.post('/favourites', (req, res) => {
     // console.log(req.body)
-    db.addCity(req.body.userId, req.body.city)
-        .catch(err => {
-            console.log(err.message)
-        })
+    if(req.body.userId && req.body.city)
+        db.addCity(req.body.userId, req.body.city)
+            .then(_ => {
+                res.status(200)
+            })
+            .catch(err => {
+                res.status(500)
+                console.log(err.message)
+            })
+    else
+        res.status(404)
 })
 
 app.delete('/favourites', (req, res) => {
     // console.log(req.body)
-    db.removeCity(req.body.userId, req.body.city)
-        .catch(err => {
-            console.log(err.message)
-        })
+    if(req.body.userId && req.body.city)
+        db.removeCity(req.body.userId, req.body.city)
+            .then(_ => {
+                res.status(200)
+            })
+            .catch(err => {
+                res.status(500)
+                console.log(err.message)
+            })
+    else
+        res.status(404)
 })
 
 app.listen(8888)
